requirements.txt
aiohttp==3.9.5
requests>=2.31.0
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.6.1
python-dotenv==1.0.1
FlagEmbedding==1.2.11
torch==2.6.0
transformers==4.57.6
tokenizers==0.22.2
cryptography>=42.0.0
py-clob-client>=0.34.6
kalshi-python-sync>=2.0.0